pyproject.toml
[project]
name = "uipath-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
]
